Initial implementation of macOS offscreen

+++ src/Avalonia.Controls.WebView.Core/Macios/Interop/CGRect.cs
@@ -1,4 +1,7 @@
 namespace Avalonia.Controls.Macios.Interop;
 
-internal record struct CGRect(float X, float Y, float Width, float Height);
-internal record struct CGSize(float Width, float Height);
+// CGFloat is double on 64-bit Apple platforms (macOS and iOS are 64-bit only).
+// Using float here causes ABI corruption: float struct fields go in s-registers,
+// but ObjC expects doubles in d-registers on arm64.
+internal record struct CGRect(double X, double Y, double Width, double Height);
+internal record struct CGSize(double Width, double Height);
